move split_headers_and_images into own mod
also add link to master screenshot array in README.md
